requirements.txt
pyusb==1.2.1
influxdb==5.3.1
influxdb-client==1.36.1
python-pushover==0.4
requests==2.25.1
